requirements.txt
python-telegram-bot
httpx
chromadb==0.5.23
sentence-transformers==2.3.1
google-generativeai==0.8.3
gspread==6.0.2
google-auth==2.28.0
google-auth-oauthlib==1.2.0
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
twilio==8.12.0
 